refactor: change structure code
