+++ src/components/shared/dropdown.type.ts
@@ -1,0 +1,14 @@
+export interface DropdownOption {
+  label: string;
+  value: string;
+}
+
+export interface CustomDropdownProps {
+  options: DropdownOption[];
+  value?: string;
+  onChange: (value: string) => void;
+  placeholder?: string;
+  isDisabled?: boolean;
+  className?: string;
+  'aria-label'?: string;
+}
